requirements.txt
uvicorn==0.12.3
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
fastapi==0.88.0
h5py==3.7.0
imutils==0.5.4
keras==2.10.0
mlflow==1.29.0
numpy==1.23.3
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
pydantic==1.10.2
pytest==7.2.0
python-dotenv==0.21.0
scikit_image==0.19.3
scikit_learn==1.1.3
setuptools==63.2.0
tensorflow==2.10.0
cmake
# dlib==19.22 this is done in dockerfile 